+++ src/main/kotlin/net/aksingh/owmjapis/model/param/Wind.kt
@@ -24,13 +24,13 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
   @field:SerializedName("speed")
-  val speed: Float? = null,
+  val speed: Double? = null,
 
   @field:SerializedName("deg")
-  val degree: Float? = null,
+  val degree: Double? = null,
 
   @field:SerializedName("gust")
-  val gust: Float? = null
+  val gust: Double? = null
 ) {
 
   fun hasSpeed(): Boolean = speed != null
